requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
DateTime==4.2
hashlib==20081119
idna==2.6
mysql==0.0.1
MySQL-python==1.2.5
pytz==2017.2
requests==2.18.4
urllib3==1.22
zope.interface==4.4.2
Jinja2==2.9.6